requirements.txt
fastapi 
uvicorn 
pdfplumber 
python-docx 
transformers 
sentence-transformers 
milvus 
streamlit 
numpy 
scipy 
pytest
python-multipart
